Prevent loading gateway pages in frames
And remove a reference to 'top.' which we won't need when we know
we're always in the top frame.

The result switcher's allowClickjacking call will override the
preventClickjacking call when we DO expect to be loaded in a frame
(i.e. when we come back from a payment processor page that we have
loaded inside a frame).

Note that merge & deployment of this patch should be coordinated
with the banner teams - they'll have to load payments-wiki in a
new window when they detect wikipedia is in a frame.

Bug: T240911
Change-Id: I2153cf92d2eabedcccae91f2c2793d81fa41c7ad