Adyen: use StagingHelper for allowedMethods
We want to minimize the risk of throwing exceptions in the
constructor, so our ResultSwitcher can use its slightly forgiving
logic to determine when to send people to the TY page.

The 'values' section of the transactions array is generally meant
for constants, not for things that take a lot of figuring out.

This patch renames the existing AdyenMethodCodec to ...Submethod...
and adds a new AdyenMethodCodec which deals with payment_method.

Change-Id: Iedb15114dba608317ffc369650a0c4be2f3ff981