Don't clear sole submethod on validate fail
Our validate code was unconditionally clearing the payment_submethod
selection when showing a form validation error. On card forms where
multiple submethod radio buttons are present, they act as submit
buttons and this is meant to leave the radio buttons unselected so
that the donor can click them again to continue. However, on iDEAL
forms where the radio button is hidden and we show a continue button
instead, we should not clear the hidden, checked payment_submethod.

This checks to ensure that multiple options exist before clearing
the payment_submethod checkbox on form validation error (and also
when clearing on page load, though that might not be a problem).

The call to clear the submethod is also moved out of the bowels
of the validate_personal function and pushed up to where we get
the results of the overall validation. We only want to clear the
selection if the validation has been triggered by selecting a
payment_submethod, not if it's been triggered by clicking a
continue button.

Bug: T251025
Change-Id: I12a3e95c2c1c147b3d44dd1191149975ee65613e