Explode out Adyen form settings by country
FIXME: we should do https://phabricator.wikimedia.org/T131989

This better matches the silly existing logic of the form chooser,
representing exactly which options are available for each country
so the chooser doesn't pick adyen over ingenico for US. It was
previously doing that (despite the low selection_weight) beacuse
of the multitude of payment submethods listed in the form settings
for adyen-usd.

Also alphebetizes by country.

The test form with selection weight zero should mean we can still
steer donors to Adyen in a variety of countries.

Bug: T254274
Change-Id: I44d0e341dacee731b43d994a05dbbdbc0677250c