Sic Procrustes on transactionSpecificValues
We don't pad them out to length like the rogue smith might have, but
this does at least cut them down to the lengths specified in
data_constraints.yaml. Previously that was only applied to staged_data.

Fixes a test that fails after truncation because it used an invalidly
long API login name.

Bug: T261058
Change-Id: If0d9681ffe2db59115b78811cbdbc9e178409deb