Workaround for Adyen payment submethod not found
The visa-debit value a legitimate submethod, but we don't actually
need to show a separate button for it. We just need it to be in the
list of available submethods. By only making it available for country
code XX we hide it from the forms while avoiding throwing an exception
in gatewayAdapter::getPaymentSubmethodMeta.

Bug: T264941
Change-Id: I55dde7e42c939926308dfc844265c93731434bca