Allow longer order IDs for Ingenico
We seem to have been confused between their optional 10 digit integer
field merchantOrderId and their alphanumeric field merchantReference,
which is what this order_id actually gets mapped to. Their docs say
it's a 40 char field, but a note says 'the maximum length of this
field for transactions processed on the GlobalCollect platform is 30.'
so this patch uses the lower of the two.

Bug: T274301
Change-Id: Ia6d07aea3b40a3eeb19a207d1eb8ca1e0aa2dec8