Add response data consistently to both data copies
Ugh, so we are storing the normalized data both in $this->dataObj
and $this->unstaged_data. The addResponseData function tries to
preserve all the original normalized data in the dataObj copy and
only overwrite the specified keys. However, it doesn't apply the
same protection to $this->unstaged_data, which is completely
overwritten from $this->staged_data.

This patch preserves all keys in $this->unstaged_data except the
keys specified in the addResponseData call. This keeps the two
data copies more synchronized, and lets us keep data from the
querystring that may be lost on the staging -> unstaging round
trip.

We also fix some places that were depending on the incorrect
behavior, updating them to explicitly state which keys they are
expecting to overwrite. In the process we get rid of the non-normal
payment_product_id key from the list of 'normalized' keys in the
Ingenico adapter, replacing it with the payment_submethod key that
it actually maps to.

Bug: T199680
Change-Id: Ie282a2dc61a0c18b3340b3c1f0cdb4eb82de67b0