Adyen Checkout: basic encrypted card submission
Uses the 'donate' mediawiki API action defined in donation.api.php
to submit our form fields along with the encrypted parameters from
the Adyen-rendered fields. The standard forms.js callDonateApi fn
is extended to allow passing in extra data.

Uses the existing SmashPig approvePayment call to capture payments.
Note that approvePayment is currently implemented as a SOAP call,
but it appears we can mix and match.

TODO: handle errors, add a new param to donation.api.php for the
encryptedExpiryYear (or just a whole blob of encryptedPaymentData)

Depends on SmashPig patch noted below.

Bug: T281528
Depends-On: I5bfb3405a072ed81ebe8e06697bc5c45646a1f06
Change-Id: Iba8c60f974c0b1d6e33d53fe20bbb434e59f20f4