Adyen Checkout: better define success + fail paths
This shows the failure message if either the auth fails or our
filter scores fall in the 'reject' range. This also finalizes the
transaction status with the correct value (not always complete)
and gets rid of an early return. Handle redirects to the TY page
in the JS.

FIXME: the existing donation.api.php code assumes that successful
doPayment means we are going to redirect the user, and logs the
pending message. Here we're doing that before the auth and capture
so as to have some useful data in the logs in case those fail, so
the extra logging from donation.api.php should be skipped.

Change-Id: I194e037d9c3e5e75e7960c731805e1cb12436fc3