QS option for debugging monthly convert
Checks for a 'debugMonthlyConvert' boolean on the querystring to
force rendering of the monthly convert modal in Mustache without
first making a donation. Note that this flag does NOT load the
JS or CSS - those still depend on the variant=monthlyConvert flag.
So in general you will want to paste all this on the end of a URL:

  &debug=true&debugMonthlyConvert=true&variant=monthlyConvert

Bug: T283262
Change-Id: I3284e82d861b5276f73a6141c37094e73f0fcbec