Adyen Checkout: first step of 3D Secure (Redirect)
So far this gets us redirected to the test challenge page and back
to the Checkout ResultSwitcher, but the checkout resultswitcher is
not doing the capture - that happens in the next patch.

Note that this only sends the browser info needed to trigger 3D
Secure rules when the country and currency have 3D Secure enabled.
e.g.
$wgDonationInterface3DSRules['USD'] = ['US'];

Need to figure out the best way to pull in the 5 JS-derived device
fingerprinting bits. Looks like we may want to create an Adyen-
specific MW API action that accepts more parameters. For now we
send placeholder values that work to trigger the 3D Secure flow.

Bug: T282730
Change-Id: I8d1b7a234f967b38fe987a01c7edf7c8ef131a6b