Adyen Checkout: handle return from redirect
Uses info returned from payment details API call to finish
processing payments. This has been tested to work with the 3D
Secure redirect flow, but might also work for iDEAL.

Breaks the doPayment function into two halves, since we need
to do basically the same stuff on return as we would do in the
second half of the single-page flow.

Bug: T282730
Change-Id: Ia2b4bc8d78b584ff039f6940cd2e836f83463ce5
Depends-On: Id98c5ab03dd0ec51409d6d42348ee831a653f184