Adyen Checkout: Add gateway_txn_id earlier
So it's present in logs and for the antifraud queue message. This
will be the txn ID of the authorization, not the capture, but if
we happen to stop the transaction because of fraud filters, we
can use the auth PSP reference to look up the payment in the
Adyen merchant console.

Bug: T282641
Change-Id: I76c738d1777ddcb69fd40b264cd764de7ad6c6cd