Configurable default monthly convert module
When monthly convert is turned on by country, we bypass the config/
variant mechanism and just inject a specifc ResourceLoader module
directly in Mustache.php's getResources. This patch allow you to
configure which of those is used by default.

Note that before this patch, we were ALWAYS adding the base
ext.donationInterface.monthlyConvert module in
Mustache::getResources. In practice, this never conflicted with
different variants of monthlyConvert because all of those variants
depended on that base module. Now that we can configure the default
to be e.g. ext.donationInterface.monthlyConvert_011 we don't want
to apply that if a different monthly convert module is called for
by a variant explicitly specified on the query string.

These are the settings I used to test locally:
$wgDonationInterfaceMonthlyConvertCountries = ['US', 'JP'];
$wgDonationInterfaceMonthlyConvertDefaultModule = 'ext.donationInterface.monthlyConvert_011';

Bug: T287262
Change-Id: I57fbe105c926bd82744d0230fb5968cf5bd97df8