Get contact info from Apple Pay sheet
Now using Promise.all to wait for two callbacks to be called
instead of relying on the order, but I had to wrap a lottt of code
inside the promise constructor to do so. FIXME: unwrap that

Also overrides the getRequiredFields function so that Apple Pay
users never see entry blanks for information we can get from
the Apple Pay component, regardless of what's in country_fields.

Bug: T285722
Change-Id: I4e716ee872b4cdf8f52d653d15703efa84e71d4b