Use Apple-branded 'donate' button
I tried this out to see if it shows anything different in the pay
sheet, but it doesn't - the line there still says 'Pay WikiMedia'.

The button seems to localize in the majority languages of the
countries where Apple Pay is offered.

We also remove an extra validate() call since we've already got it
up at the start of the onSumbit handler

Change-Id: If6ec57e0521bf368999b989c0713240295ec2c65