Fix employer select arrow key navigation
Show name in select box, not ID. Pressing down triggered the menufocus
function in jq.autocomplete, which set the value to item.value when
the item was a keypress and the focus callback was not handled.

This patch just handles the focus callback with the same function that
handles the select callback.

Bug: T292582
Change-Id: Id24044f3536c689ca686331bd45ad08a61b0c27f