Revert "Adyen Checkout: Switch ApplePay logic to use shipping contact by default"
This reverts commit 1153abed2f59befded9cb0cfe12b8e34164c47b5.

Reason for revert: We don't want to push this out just yet. David asked us to have the patch at the ready but it's not yet been decided if we're going to switch over. Let's keep it on hold to save the cherry-pick dance each time we push out payments. 

I also share your reservations about how useful it'll be if I interpreted that correctly. 

Change-Id: I7acee76e2f380d9b494aae84791bcfec4440a3bc