Move Monthly Convert ask rules to config
Also, only send the rules for the selected currency, as that's known
server-side. Slightly simplifies the JS. With the default settings,
EUR still falls back to GBP rules and NZD/CAD/AUD still fall back
to USD rules, but it is now possible to add specific rules for each
of those currencies (or any other currency).

Note that the [ Infinity, 0 ] bits at the end of the old rules
are neither compatible with JSON nor necessary - if we get to the
end of the rules array and haven't found an entry where the first
value is greater than the given amount, getConvertAsk returns 0.

Bug: T283834
Change-Id: Ic8eb0a9dabe7e6496710848bc68c0d0d267029b5