Short-circuit Adyen JS on ResultSwitcher
Looks like the setup function isn't actually needed for anything
when we're loading the Adyen ResultSwitcher page - at that point
we've done all we want to do with the Adyen Checkout components.

Note that this page is only shown when a donor is using a
3DSecure card and we are ALSO showing the Monthly Convert form.

This patch also adds a showSubmethodButtons override to the Adyen
ResultSwitcher page so that the background form doesn't have
blank submethod boxes.

Test with $wgDonationInterfaceMonthlyConvertCountries set to an
array including your favorite test country and with a card from
https://docs.adyen.com/development-resources/testing/test-card-numbers#test-3d-secure-2-authentication

Bug: T307485
Change-Id: I03bd18bbf15afd29e06bfa01d3b1b6a4c3f9b9e9