Remove PayPal legacy adapter
And update/delete a couple of tests that used it. The redirection
logging for remaining adapters generally happens on API submission,
not on gateway page submission. PayPalApiTest::testGoodSubmit
already verifies the redirection logging for that case, so we are OK
to delete the other test.

For the 'IntegrationTest', the preconditions being tested were
particular to the Paypal legacy adapter (ffname pushing is also
completely obsolete at this point). Replaced with a check that the
contribution tracking ID is being regenerated, which we expect to
happen when switching gateways.

Bug: T300068
Change-Id: Ia6b9466be805276ed3293e004906f2bf92952201