Simpler function to build payments form URLs
If we need to access this statically later we can add another
param to take the Config object instead of using $this->getConfig.

Adds assertion that the built URL contains the passed-in parameters,
updates the test recurring param to strings so assertArraySubmapSame
is happy. Technically more correct I guess, as everything on the QS
is a string.

Update canonical params test to check for special page title rather
than for ffname as we are phasing out ffname.

Bug: T302944
Change-Id: If12bee5e6fb6c4c5bec10f69b81d68ba7cbd1007