Re-hide extra Visa logo
It was being filtered out with a country code list that only included
'XX'. We now have tests that make sure all the countries in the
submethod lists are supported by the whole gateway, so we can't use
XX. We can still filter it out by using the empty list.

Later we should change the decoder logic to use SmashPig's
ReferenceData as we do in Adyen and AstroPay, but this seems like the
least risky fix for right now.

Change-Id: I5f70884e8c832e661505d504fa2ab7c24c34e1f6