Move showError logic up to GatewayPage
This way we don't need to stash showError in the DonationData object
and we don't need the getFormClass function. Keeps concerns slightly
better separated.

Also only registers the MakeGlobalVariablesScript hook when we're
rendering a real form, not an error form. This way we don't waste
API calls to retrieve client tokens that are only useful on non-
error forms.

Bug: T302939
Change-Id: I4a3646bdd00c4906a818b6824c262518eb53fc45