Clean up Ingenico config
We only support cc and rtbt through the current adapter.
The previous adapter supported other things at some point, but
e.g. the dd config wouldn't even make sense with the current
adapter's code as it indicates XML API tag names.

Keys additional_success_status and short_circuit_at also are only
referred to in a GlobalCollect-specific function where they are
used to map different codes to success based on the submethod.
For the Ingenico adapter that logic is encapsulated in the
different SmashPig PaymentProvider classes.

Bug: T302939
Change-Id: Iae0d0bd0fb5fd85d8372517aef128c4d7b4252df