3DS: move card type check to Ingenico subclass
The base canSet method now just checks for the minimum necessary
information to call the isRecommend method.

For Adyen we can send the extra browserInfo whenever the currency/
country rule tells us to. If we send it for an unsupported card
type it will just be ignored. Under Ingenico, we would get errors
if we tried to enable 3DSecure on an unsupported card type.

Also updates the currently-unused-in-production GlobalCollect class,
just to keep tests running till we get around to deleting all the
GlobalCollect things.

Bug: T313756
Change-Id: Idc6125f4f617482b097689273449133edfba9ed6