Adyen: send 'card on file' for MC tokenization
When we're not sure if we will use the token again, use the new
recurring_model parameter to indicate that this should be tokenized
as a 'card on file' transaction rather than a subscription. Adyen
has indicated that we can later use the token with the 'subscription'
recurring model if they do add a monthly donation.

Bug: T313854
Change-Id: I8f7bfaab3437004bd5942162f41f9d4e3a65a384