GatewayChooser: fix submethod support check
Got a little overzealous with the in_array() use in the previous patch
(I253f45088c2cd37e18d7d8b6c1bfd96ef5b04163).

in_array() should only be used to check array values, however, in the
previous patch, it was being used in two locations for structures
where we needed to check array keys instead. This patch corrects
that.

(The confusion was due in part to the inconsistency between
how country restrictions are set at the submethod level and how
they are set elsewhere. At submethod-level, country restrictions
are not a flat array of country codes like the lists for gateway
and method-level restrictions, but an associativ array of
country code => bool. This patch also adds a FIXME for this issue.)

Bug: T315899
Change-Id: I6e171963d5bfa53608c11b576e5dec7f85c99812