Tweak log message to avoid possible error
Probably won't happen in production, but in running local tests
with the MonthlyConvertCountries set to true, I saw a few errors
where the 'gateway' key wasn't set when we hit this line.

Since log lines are all prefixed by the gateway name, this info
is redundant to put in the message.

Change-Id: I15269752cdbe8693cace0902ef4b345b7e8eaa77