Catch errors in setClientVariables
That MakeGlobalVariablesScript hook is called really late in the
page lifecycle, after we've already rendered the form and finalized
the headers, so we can't replace the HTML with a fail page or even
send a Location header to redirect to one. Instead, we just set
$vars that instruct the base forms JS to do the redirect in the
browser.

We don't want to just suppress the error as that will lead to
mysterious form malfunctions when the expected variables are not
present.

I tested this by editing PaymentProvider::getPaymentMethods in
in payments/vendor/wikimedia/smash-pig/PaymentProviders/Adyen to
make it immediately throw an exception.

Bug: T295149
Change-Id: Iac84b58f655ada402e9f129f81880e3be50a38a4