First variant for Japan tests
Changes donation amount, billing info, and placeholders for fname,
lname, and expiration date.

https://docs.google.com/presentation/d/1d1QkcJTj5MEwE201BnJ03DpdL4_zefoT7joLBAOnE0Q/

To change the expiration date we have to feed different parameters
to the Adyen Checkout constructor, so we need to run this variant
script BEFORE the adyen script is run. That's why we have the ugly
copy-pasted adyenCheckoutWithJaVariant01A module - we need another
version of the adyenCheckout module with the variant script as a
dependency. Note that we cannot programatically change the
expiration date placeholder after the widget is created because
the input is created in an iframe hosted on Adyen's site (to give
us PCI-compliant separation).

Test with &variant=ja_01A on a JP/ja Adyen checkout form link e.g.
https://localhost:9001/index.php?title=Special:AdyenCheckoutGateway&appeal=JimmyQuote&payment_method=cc&recurring=0&language=ja&uselang=ja&country=JP&currency=JPY&amount=150&variant=ja_01A

Bug: T322793
Change-Id: Ic335f90cf981fcc36cfa37d5594b54602d024714