Ingenico disentanglement: clean up doPayment
Use the normalized createPaymentSession and the result properties,
instead of building the whole nested request in DonationInterface
and parsing the raw response. Copy and modify the old ReturnUrl
helper so we can just delete the old one along with its old tests
later.

Since we modify var_map and transformers here, we introduce
temporary backup copies of the legacy versions, which are restored
at the start of the legacy do_transaction call.

The unit tests have been changed to integration tests, to make
sure we are sending the same parameters to Ingenico and not losing
anything in the SmashPig / DonationInterface interaction. As a
result we update a couple of expectations to e.g. look for 3ds
indicators and the recurring sequence indicators in the new place
(where current docs indicate they should be) and we change the
type of a couple boolean flags from string to bool or vice versa.
Fortunately ingenico seems to have been putting up with the
mixed strings and booleans thus far.

Note that this depends on one more SmashPig patch which fixes the
Monthly Convert behavior, sending different parameters to get a
token.

Bug: T260725
Change-Id: Ib00c01f244b7525811476222e722bfaf4beea2b1