Ingenico disentanglement: modern processDonorReturn
In this patch we take advantage of the pre-parsed properties of the
getLatestPaymentStatus and approvePayment API calls rather than
re-parsing everything in DonationInterface.

We remove the status lookup and approvePayment from the transactions
metadata array because we only use that for building the params
now, and those have very simple parameters.

We also move the Ingenico-specific log sanitization out of the parent
adapter class and delete the Ingenico CVV and AVS maps since those
scores are now calculated in the SmashPig code.

Next patch can remove lots of now-unused code.

Test changes:

* instead of individual do_transaction, do the full processDonorReturn

* removed testGetHostedPaymentStatus25 as it tested a scenario that
would throw an exception in a non-mocked HostedCheckoutProvider
and apparently hasn't happened in years. The 'APPROVED' status
doesn't seem to happen for cards and isn't in the enum list of
statuses for the hostedcheckout status lookup

* testApprovePayment removed, assertion added to testDonorReturnSuccess

Bug: T260725
Change-Id: Ic05c9462277a7df835749b511256b6d3fab51c10