Break up payments_form and form_amount in c_t messages
Needs consumer patch deployed first (see Depends-On).

This also adds individual components for the payment method,
submethod, and recurring-ness of the transaction (though it leaves
them in the concatenated version of utm_source for now).

Also alphabetizes the field list and removes 'anonymous' and 'note'
fields which are now ignored by the consumer.

Also sends the old concatenated versions of the fields till we are
ready to stop populating the legacy table.

Bug: T334904
Depends-On: I268b1d5633494bb817c7112f460479a9f3ac8def
Change-Id: I2f29f680a3bd3fe0b6a75b90b7f3df956efd732d