Add user-agent info to contribution_tracking
To help diagnose browser-specific issues.

The WhichBrowser library seems to be a lot more up-to-date than the
library mentioned in the ticket.
You'll be able to see them pushed to the queue with e.g.
scripts/redis-queues-monitor, but we need a CRM patch to actually
get them into the civicrm_contribution_tracking table.

Bug: T182039
Change-Id: Ia3fecd23d333f17bac9609254c3e0c8c06da90aa