Verified with sandbox and live credentials via postman
Test:
https://localhost:9001/index.php?title=Special:DlocalGateway&appeal=JimmyQuote&payment_method=bt&recurring=0&uselang=en&language=en&currency=BRL&amount=100&country=BR&first_name=Jimmy&last_name=Wales&fiscal_number=332.945.766-09&email=jwales%40example.com&debug=true
should direct to pix payment page.

Still wait for dlocal's response on Bank transfer ACH with bank code as IO, which see from their doc
https://docs.dlocal.com/docs/brazil#examples-3
but always return Merchant has no authorization to use this API, so remove ACH for latam but add za, also test za is not valid for stitch but io working...
all not tie with their documentation
in order to still support brazil bank transfer use SB which is always redirect to pix
see from email thread with title "Incorrect flow for Bank Transfers. It is leading to PIX
"

Bug: T355371
Change-Id: I29e774806cae5cc3f67e5a6a7c122791da287db2