Move $recurringOptions calculation out of EmailForm
And fall back to the USD options when the selected currency is not
configured. The original currency won't change, we just might get
something silly like asking for a 1 JPY upgrade. Also feed in the
formatted value to the form so it can be interpolated into a string,
rather than repeating the locale and currency for each option.

Bug: T356370
Change-Id: If2518daba0e0685709d5c4d15c9cb37681d5c9b5