Better amount validation for RecurUpgrade
Server-side validation that it's positive, and also disallow huge
numbers (configurable with DonationInterfaceRecurringUpgradeMaxUSD).

We should really only have an amount to validate after the page is
posted back, so we can be pretty sure we have the currency in
session.

Also fixes an error where amount validation was being skipped
entirely because I didn't understand how to use ??=

Easiest way to test the server-side validation is to damage the
client-side validation by e.g. setting maximum to 100000 in the js.

Bug: T359694
Change-Id: Ice409e90e9fbb498b458226888dd604537fd30d1