Move RecurUpgrade validation into its own class
With a fancy namespace and tests and all.

Phan seems not to know about classes that are not PSR-4 compliant
nor classes in vendor, unless they are referred to in its config.
We add the classes referred to by the Validator there so phan
doesn't fail us on unknown classes, but we also indicate that they
are not yet ready for phan to analyze them.

Change-Id: I2e2e223169c7ecdb4c544194c04c18e954c0c445