Set RecurUpgrade title with less code
The title is already being set inside setHeaders(), to the value
returned by getDescription. And getDescription just sets it to
the message matching the lowercase version of the special page
name.

In the non-variant template we were using the old title as the
header for the amount selection box, but that's misleading. Use
the recurupgrade-amount-ask message there instead.

Change-Id: Iba9a4722cb1ddfd874d4e12c1821b76974318845