Fix prefs display when contact snoozed
It was rendering the opt-in button as checked before. This fixes the
list of allowed params to pass through isSnoozed, and uses it in the
template to assure that ONLY the snoozed option is checked when true.

Also only count a contact as snoozed if their snooze_date is in the
future. Prior to this patch is was counting any date as isSnoozed.

Bug: T358878
Change-Id: I200db4c093bd3451e40f8d62cda2a4a49a3792a1