Fix Monthly Convert min amount validation
This is for when you click the 'other amount' link then put in
something tiny and click the donate button.

Oops, when we moved over to the amount rules in yaml and removed
wgDonationInterfacePriceFloor, we never updated this. The other
amount validation should now work, and will be available to validate
the other amounts for different variants.

Bug: T368155
Change-Id: Iad3cebf29c6f77c459a4c511fe8d80f4dd8ec740