Make annual recurring convert look like latest mockup
Now with more buttons, and possibility to set up other amounts for
annual or monthly donations

Instead of using message_replacement, we replace the whole monthlyConvert
partial mustache template since there are more HTML changes from the base.

Add &variant=monthlyConvert_annual to the adyen URL

Also fixes the start_date for annual recurring to be a year in the future
rather than a month.

Bug: T368155
Change-Id: Ia29ea8054c09c9d84fcdc36498b2b95935612ae8