Add missing config file to gravy gateway
This weird little config file should probably be removed from
everywhere and the gateway chooser should test for the
FallbackCurrency mw config variables, as those are the only way
we want to include the gateway when the currency given is not
on the supported list.

Bug... till we do that change, this will let tests pass locally.

Change-Id: I105389d71d24888830d40a0c1bca443ccdaeaef2