Generic velocity filter
The 'GenericVelocityFilter' class is a stripped down version of the
IP velocity filter, without the allow / deny lists or the variable
fail score.

The 'VelocityFilterRunner' class is responsible for reading the
config and instantiating all the configured filters, and feeding
them the cache and session objects they need to run.

They are wired in to the old-fashioned custom_filters fraud filter,
though their config is a bit different.

Test config is checked in to config-private.

Bug: T221378
Change-Id: I87e8d24ec9310adc047d21ad0d4aae163917effc