Add properties for DonorPortal amount validation
Adds the currency rates and the donation rules for each specific
recurring contribution to the JS, so we can use the same logic as
MonthlyConvert does to validate amounts.

Since we're re-reading the whole config for each rule lookup, I
limited the lookups to only active donations with can_modify=true.
Note that should still include paused.

Bug: T388753
Bug: T388755
Depends-On: I79da9429eeb01d179cb37f1e135d4c22f4b8c073
Change-Id: I4fe950a64978afa289c16e6e3c1b026c71e90083