Allow Venmo 'newBrowserTab' session.
From the docs:
" If options.allowNewBrowserTab is false when calling venmo.create,
this method will return true only for browsers known to support
returning from the Venmo app to the same browser tab.
Currently, this is limited to iOS Safari and Android Chrome. "

https://braintree.github.io/braintree-web/3.86.0/Venmo.html#isBrowserSupported

We're not sure why we'd want to restrict payments to a single tab,
so let's turn it off in case it's causing the issues we're experiencing elsewhere.

Change-Id: Ic673d34d605eca46a5aed9bd38a94da6e9fb3972