Unlink other amount box from other radio buttons
On the donor portal form, try to cancel a monthly donation, then
select 'switch to an annual gift'. Click on the 'other amount'
radio button and enter some value. Then click on the other radio
buttons and confirm that the other amount box does not change.

The confirmation text at the bottom and the final submitted amount
should follow the appropriate selected radio button.

Bug: T413106
Change-Id: I88c0a0ec7dcecb04f2a731c24e770c34a4d55c2e