Venmo: pass totalAmount at click time
Pass totalAmount when creating the Venmo component so the amount is
shown on the Venmo paysheet. The amount is read at click time to avoid
stale values after donor edits or validation errors.

Braintree JS SDK v3.96.1 and later requires the 'totalAmount' parameter
to be passed during Venmo component creation if the amount should be
displayed to the user on the Venmo paysheet. We recently bumped the
Braintree JS version from v3.96.0 to v3.133.0, making this requirement
active.

https://developer.paypal.com/braintree/docs/guides/venmo/client-side/javascript/v3/#amounts-and-line-items

Alternative to https://gerrit.wikimedia.org/r/c/mediawiki/extensions/DonationInterface/+/1219580

Bug: T413074
Change-Id: I2a0486919038ad4d1aaa85a30468809e665870ff